requirements.txt
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
coverage==7.10.7
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.31.0
idna==3.10
iniconfig==2.1.0
oauthlib==3.3.1
packaging==25.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
